split search.cc into small files
