+++ src/search/search.cc
@@ -1,7 +1,9 @@
 #include "aso/search/search.h"
 #include "aso/kernel/customized.h"
+#include "aso/search/dim_strategy.h"
+#include "aso/search/op_utils.h"
+#include "aso/search/tensor_utils.h"
 #include "aso/utils/containers.h"
-#include "aso/utils/hash_utils.h"
 
 namespace aso {
 namespace search {
@@ -12,366 +14,6 @@ std::vector<int> to_dim_vector(int num_dims, int *dim) {
     dims.push_back(dim[i]);
   }
   return dims;
-}
-
-template <typename Op, typename OpType>
-size_t get_operator_hash(Op i, OpType op) {
-  size_t h = 0;
-  hash_combine(h, i);
-  hash_combine(h, op);
-  return h;
-}
-
-template <typename Op, typename OpType>
-size_t get_operator_hash(Op i, Op j, OpType op) {
-  size_t h = 0;
-  hash_combine(h, i);
-  hash_combine(h, j);
-  hash_combine(h, op);
-  return h;
-}
-
-bool is_binary(type::TBOperatorType op) {
-  std::unordered_set<type::TBOperatorType> true_values{
-      type::TBOperatorType::TB_MATMUL_OP, type::TBOperatorType::TB_DIV_OP};
-  return contains(true_values, op);
-}
-
-bool is_unary(type::TBOperatorType op) {
-  std::unordered_set<type::TBOperatorType> true_values{
-      type::TBOperatorType::TB_EXP_OP,
-      type::TBOperatorType::TB_REDUCTION_0_OP,
-      type::TBOperatorType::TB_REDUCTION_1_OP,
-      type::TBOperatorType::TB_REDUCTION_2_OP};
-  return contains(true_values, op);
-}
-
-bool is_binary(type::KNOperatorType op) {
-  std::unordered_set<type::KNOperatorType> true_values{
-      type::KNOperatorType::KN_MATMUL_OP};
-  return contains(true_values, op);
-}
-
-bool is_unary(type::KNOperatorType op) {
-  std::unordered_set<type::KNOperatorType> true_values{
-      type::KNOperatorType::KN_REDUCTION_0_OP,
-      type::KNOperatorType::KN_REDUCTION_1_OP,
-      type::KNOperatorType::KN_REDUCTION_2_OP,
-  };
-  return contains(true_values, op);
-}
-
-std::shared_ptr<AlgebraicPattern>
-    get_pattern(type::KNOperatorType op,
-                std::shared_ptr<AlgebraicPattern> opd) {
-  switch (op) {
-    case type::KNOperatorType::KN_REDUCTION_0_OP:
-    case type::KNOperatorType::KN_REDUCTION_1_OP:
-    case type::KNOperatorType::KN_REDUCTION_2_OP:
-      return opd;
-    case type::KNOperatorType::KN_OUTPUT_OP:
-      return opd;
-    default:
-      assert(false);
-  }
-}
-
-std::shared_ptr<AlgebraicPattern>
-    get_pattern(type::TBOperatorType op,
-                std::shared_ptr<AlgebraicPattern> opd) {
-  switch (op) {
-    case type::TBOperatorType::TB_EXP_OP:
-      return std::make_shared<Exp>(opd);
-    case type::TBOperatorType::TB_REDUCTION_0_OP:
-    case type::TBOperatorType::TB_REDUCTION_1_OP:
-    case type::TBOperatorType::TB_REDUCTION_2_OP:
-      return opd;
-    case type::TBOperatorType::TB_OUTPUT_OP:
-      return opd;
-    default:
-      assert(false);
-  }
-}
-
-std::shared_ptr<AlgebraicPattern>
-    get_pattern(type::KNOperatorType op,
-                std::shared_ptr<AlgebraicPattern> lhs,
-                std::shared_ptr<AlgebraicPattern> rhs) {
-
-  switch (op) {
-    case type::KNOperatorType::KN_MATMUL_OP:
-      return std::make_shared<Mul>(lhs, rhs);
-    default:
-      assert(false);
-  }
-}
-
-std::shared_ptr<AlgebraicPattern>
-    get_pattern(type::TBOperatorType op,
-                std::shared_ptr<AlgebraicPattern> lhs,
-                std::shared_ptr<AlgebraicPattern> rhs) {
-
-  switch (op) {
-    case type::TBOperatorType::TB_MATMUL_OP:
-      return std::make_shared<Mul>(lhs, rhs);
-    case type::TBOperatorType::TB_DIV_OP:
-      return std::make_shared<Div>(lhs, rhs);
-    default:
-      assert(false);
-  }
-}
-
-size_t get_rdeg(type::KNOperatorType op,
-                int input_rdeg,
-                DTensor const &input_tensor) {
-  switch (op) {
-    case type::KNOperatorType::KN_REDUCTION_0_OP:
-      return input_rdeg * input_tensor.dim[0];
-    case type::KNOperatorType::KN_REDUCTION_1_OP:
-      return input_rdeg * input_tensor.dim[1];
-    case type::KNOperatorType::KN_REDUCTION_2_OP:
-      return input_rdeg * input_tensor.dim[2];
-    default:
-      return input_rdeg;
-  }
-}
-
-size_t get_rdeg(type::TBOperatorType op,
-                int input_rdeg,
-                STensor const &input_tensor) {
-  switch (op) {
-    case type::TBOperatorType::TB_REDUCTION_0_OP:
-      return input_rdeg * input_tensor.dim[0];
-    case type::TBOperatorType::TB_REDUCTION_1_OP:
-      return input_rdeg * input_tensor.dim[1];
-    case type::TBOperatorType::TB_REDUCTION_2_OP:
-      return input_rdeg * input_tensor.dim[2];
-    default:
-      return input_rdeg;
-  }
-}
-
-bool check_tensor_shape(type::TBOperatorType op, STensor const &input) {
-  switch (op) {
-    case type::TBOperatorType::TB_EXP_OP:
-      return true;
-    case type::TBOperatorType::TB_REDUCTION_0_OP:
-      return input.dim[0] != 1;
-    case type::TBOperatorType::TB_REDUCTION_1_OP:
-      return input.num_dims >= 2 && input.dim[1] != 1;
-    case type::TBOperatorType::TB_REDUCTION_2_OP:
-      return input.num_dims >= 3 && input.dim[2] != 1;
-    default:
-      assert(false && "Unsupported Operator");
-  }
-}
-
-bool check_tensor_shape(type::TBOperatorType op,
-                        STensor const &input1,
-                        STensor const &input2) {
-  switch (op) {
-    case type::TBOperatorType::TB_MATMUL_OP:
-      return input1.num_dims == 2 && input2.num_dims == 2 &&
-             input1.dim[1] == input2.dim[0];
-    case type::TBOperatorType::TB_DIV_OP:
-      return input1.num_dims == 2 && input2.num_dims == 2 && input2.dim[1] == 1;
-    default:
-      assert(false && "Unsupported Operator");
-  }
-}
-
-bool check_tensor_shape(type::KNOperatorType op, DTensor const &input) {
-  switch (op) {
-    case type::KNOperatorType::KN_REDUCTION_0_OP:
-      return input.dim[0] != 1;
-    case type::KNOperatorType::KN_REDUCTION_1_OP:
-      return input.num_dims >= 2 && input.dim[1] != 1;
-    case type::KNOperatorType::KN_REDUCTION_2_OP:
-      return input.num_dims >= 3 && input.dim[2] != 1;
-    default:
-      assert(false && "Unsupported Operator");
-  }
-}
-
-bool check_tensor_shape(type::KNOperatorType op,
-                        DTensor const &input1,
-                        DTensor const &input2) {
-  switch (op) {
-    case type::KNOperatorType::KN_MATMUL_OP:
-      return input1.num_dims == 2 && input2.num_dims == 2 &&
-             input1.dim[1] == input2.dim[0];
-    default:
-      assert(false && "Unsupported Operator");
-  }
-}
-
-unsigned int get_num_threadblock(dim3 const &grid_dim) {
-  return grid_dim.x * grid_dim.y * grid_dim.z;
-}
-
-std::vector<dim3> get_grid_dim_cand(std::vector<DTensor> const &tensors,
-                                    std::vector<int3> const &input_map) {
-  std::vector<dim3> results;
-  for (int x = 1;; x *= 2) {
-    for (int i = 0; i < tensors.size(); ++i) {
-      if (tensors[i].dim[input_map[i].x] % x != 0) {
-        return results;
-      }
-    }
-    for (int y = 1;; y *= 2) {
-      bool feasible = true;
-      for (int i = 0; i < tensors.size(); ++i) {
-        if (tensors[i].num_dims > 1 &&
-            tensors[i].dim[input_map[i].y] % y != 0) {
-          feasible = false;
-          break;
-        }
-      }
-      if (!feasible) {
-        break;
-      }
-      results.push_back(dim3(x, y));
-    }
-  }
-}
-
-std::vector<dim3> get_block_dim_cand(std::vector<DTensor> const &tensors,
-                                     std::vector<int3> const &input_map,
-                                     dim3 grid_dim) {
-  std::vector<dim3> results;
-  for (int x : {32, 64, 128}) {
-    for (DTensor const &tensor : tensors) {
-      assert(tensor.size() % get_num_threadblock(grid_dim) == 0);
-      int block_size = tensor.size() / get_num_threadblock(grid_dim);
-      if (block_size % x == 0) {
-        results.push_back(dim3{x, 1, 1});
-      }
-    }
-  }
-  return results;
-}
-
-bool is_all_replicate(std::vector<int3> const &input_maps) {
-  for (int3 const &input_map : input_maps) {
-    if (input_map.x != -1 || input_map.y != -1 || input_map.z != -1) {
-      return false;
-    }
-  }
-  return true;
-}
-
-void generate_input_map_cand(int num_tensors,
-                             int3 input_map,
-                             std::vector<int3> cur,
-                             std::vector<std::vector<int3>> &results) {
-  if (cur.size() == num_tensors) {
-    if (!is_all_replicate(cur)) {
-      results.push_back(cur);
-    }
-    return;
-  }
-  cur.push_back(int3{-1, -1, -1});
-  generate_input_map_cand(num_tensors, input_map, cur, results);
-  cur.pop_back();
-  cur.push_back(int3{input_map.x, -1, -1});
-  generate_input_map_cand(num_tensors, input_map, cur, results);
-  cur.pop_back();
-  cur.push_back(int3{-1, input_map.y, -1});
-  generate_input_map_cand(num_tensors, input_map, cur, results);
-  cur.pop_back();
-  cur.push_back(int3{input_map.x, input_map.y, -1});
-  generate_input_map_cand(num_tensors, input_map, cur, results);
-  cur.pop_back();
-}
-
-std::vector<std::vector<int3>>
-    get_input_map_cand(std::vector<DTensor> const &tensors) {
-  std::vector<std::vector<int3>> results;
-  // Assume two-dimentional inputs
-  // TODO: There are invalid input maps, how to prune them out?
-  for (int3 input_map : {int3{0, 1, -1}, int3{1, 0, -1}}) {
-    generate_input_map_cand(tensors.size(), input_map, {}, results);
-  }
-  return results;
-}
-
-void generate_forloop_dim(int num_tensors,
-                          std::vector<int> cur,
-                          std::vector<std::vector<int>> &results) {
-  if (cur.size() == num_tensors) {
-    bool is_none = true;
-    for (int dim : cur) {
-      if (dim != -1) {
-        is_none = false;
-        break;
-      }
-    }
-    if (!is_none) {
-      results.push_back(cur);
-    }
-    return;
-  }
-
-  for (int dim = -1; dim <= 2; ++dim) {
-    cur.push_back(dim);
-    generate_forloop_dim(num_tensors, cur, results);
-    cur.pop_back();
-  }
-}
-
-std::vector<std::vector<int>>
-    get_forloop_dim_cand(std::vector<DTensor> const &input_tensers) {
-  std::vector<std::vector<int>> results;
-  generate_forloop_dim(input_tensers.size(), {}, results);
-  return results;
-}
-
-std::vector<int>
-    get_forloop_range_cand(std::vector<DTensor> const &input_tensors,
-                           std::vector<int3> const &input_map,
-                           dim3 grid_dim,
-                           dim3 block_dim,
-                           std::vector<int> const &forloop_dim) {
-  std::vector<int> results;
-
-  for (int x = 1;; x *= 2) {
-    for (int i = 0; i < input_tensors.size(); ++i) {
-      if (forloop_dim[i] == -1) {
-        continue;
-      }
-      int dim;
-      switch (forloop_dim[i]) {
-        case 0:
-          dim = input_tensors[i].dim[input_map[i].x];
-          assert(dim % grid_dim.x == 0);
-          dim /= grid_dim.x;
-          assert(dim % block_dim.x == 0);
-          dim /= block_dim.x;
-          break;
-        case 1:
-          dim = input_tensors[i].dim[input_map[i].y];
-          assert(dim % grid_dim.y == 0);
-          dim /= grid_dim.y;
-          assert(dim % block_dim.y == 0);
-          dim /= block_dim.y;
-          break;
-        case 2:
-          dim = input_tensors[i].dim[input_map[i].z];
-          assert(dim % grid_dim.z == 0);
-          dim /= grid_dim.z;
-          assert(dim % block_dim.z == 0);
-          dim /= block_dim.z;
-          break;
-        default:
-          assert(false);
-      }
-      if (dim % x == 0) {
-        results.push_back(x);
-      } else {
-        return results;
-      }
-    }
-  }
 }
 
 bool KernelGraphGenerator::is_finished_graph(
@@ -718,7 +360,8 @@ void KernelGraphGenerator::generate_next_kernel(
         for (dim3 grid_dim : get_grid_dim_cand(input_tensors, input_map)) {
           for (dim3 block_dim :
                get_block_dim_cand(input_tensors, input_map, grid_dim)) {
-            for (std::vector<int> const &forloop_dim : get_forloop_dim_cand(input_tensors)) {
+            for (std::vector<int> const &forloop_dim :
+                 get_forloop_dim_cand(input_tensors)) {
               for (int forloop_range : get_forloop_range_cand(input_tensors,
                                                               input_map,
                                                               grid_dim,
